backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
email-validator==2.1.0
jinja2==3.1.2
sqlalchemy
pymysql
fastapi-mail
python-dotenv